expose the underlying handle
While using this library solely to manage lock file (e.g. foo.txt.lock
to lock foo.txt) is quite nice, it doesn't work so well when there are
external requirements to directly lock a file -- for example, when
integrating with a system with which you have no control.

I recently ran into a situation where I needed the backing file handle
in order to reliably interoperate with a system running on Windows.
Despite the limitation of the file handle not being exposed, this
library remains the best library for os-independent file locking for
golang. Extend it to support retrieving the file handle so that users
who have no other choice can still make use of the library's core
functionality.